+++ Character/Class/Warrior/Spells/Whirlwind.cpp
@@ -1,11 +1,12 @@
-
 #include "Whirlwind.h"
-#include "CooldownReady.h"
-#include "Warrior.h"
-#include "Flurry.h"
-#include "DeepWounds.h"
-#include "OverpowerBuff.h"
+
 #include "CharacterStats.h"
+#include "CombatRoll.h"
+#include "CooldownReady.h"
+#include "DeepWounds.h"
+#include "Flurry.h"
+#include "OverpowerBuff.h"
+#include "Warrior.h"
 
 Whirlwind::Whirlwind(Character* pchar) :
     Spell("Whirlwind", "Assets/warrior/Ability_whirlwind.png", pchar, RestrictedByGcd::Yes, 10.0, ResourceType::Rage, 25),
@@ -24,18 +25,18 @@ void Whirlwind::spell_effect() {
 
     if (result == PhysicalAttackResult::MISS) {
         increment_miss();
-        warr->lose_rage(resource_cost);
+        warr->lose_rage(static_cast<unsigned>(resource_cost));
         return;
     }
     if (result == PhysicalAttackResult::DODGE) {
         increment_dodge();
         warr->get_overpower_buff()->apply_buff();
-        warr->lose_rage(static_cast<int>(round(resource_cost * 0.25)));
+        warr->lose_rage(static_cast<unsigned>(round(resource_cost * 0.25)));
         return;
     }
     if (result == PhysicalAttackResult::PARRY) {
         increment_parry();
-        warr->lose_rage(static_cast<int>(round(resource_cost * 0.25)));
+        warr->lose_rage(static_cast<unsigned>(round(resource_cost * 0.25)));
         return;
     }
 
@@ -51,5 +52,5 @@ void Whirlwind::spell_effect() {
         add_hit_dmg(static_cast<int>(round(damage_dealt)), resource_cost, pchar->global_cooldown());
     }
 
-    warr->lose_rage(resource_cost);
+    warr->lose_rage(static_cast<unsigned>(resource_cost));
 }
